+++ Win32GUI/MultiEditText.h
@@ -10,6 +10,6 @@ public:
 	MultiEditText(Control*, TSTRING, int, int = 23);
 	MultiEditText(Control*, TSTRING, DWORD, RECT);
 
-	void heightInLines(int);
+	void heightInLines(LONG);
 };
